Make DependencyInjection Builder as nested class
