+++ sdk/src/main/java/com/qonversion/android/sdk/internal/di/DependencyInjection.kt
@@ -1,11 +1,18 @@
 package com.qonversion.android.sdk.internal.di
 
+import android.app.Application
 import com.qonversion.android.sdk.internal.di.controllers.ControllersAssembly
+import com.qonversion.android.sdk.internal.di.controllers.ControllersAssemblyImpl
 import com.qonversion.android.sdk.internal.di.mappers.MappersAssembly
+import com.qonversion.android.sdk.internal.di.mappers.MappersAssemblyImpl
 import com.qonversion.android.sdk.internal.di.misc.MiscAssembly
+import com.qonversion.android.sdk.internal.di.misc.MiscAssemblyImpl
 import com.qonversion.android.sdk.internal.di.network.NetworkAssembly
+import com.qonversion.android.sdk.internal.di.network.NetworkAssemblyImpl
 import com.qonversion.android.sdk.internal.di.services.ServicesAssembly
+import com.qonversion.android.sdk.internal.di.services.ServicesAssemblyImpl
 import com.qonversion.android.sdk.internal.di.storage.StorageAssembly
+import com.qonversion.android.sdk.internal.di.storage.StorageAssemblyImpl
 
 internal class DependencyInjection(
     private val mappersAssembly: MappersAssembly,
@@ -19,4 +26,25 @@ internal class DependencyInjection(
     MiscAssembly by miscAssembly,
     StorageAssembly by storageAssembly,
     NetworkAssembly by networkAssembly,
-    MappersAssembly by mappersAssembly
+    MappersAssembly by mappersAssembly {
+
+    class Builder(private val application: Application) {
+        fun build(): DependencyInjection {
+            val mappersAssembly = MappersAssemblyImpl()
+            val miscAssembly = MiscAssemblyImpl(application)
+            val storageAssembly = StorageAssemblyImpl(mappersAssembly, miscAssembly)
+            val networkAssembly = NetworkAssemblyImpl(mappersAssembly, storageAssembly, miscAssembly)
+            val servicesAssembly = ServicesAssemblyImpl(mappersAssembly, storageAssembly, networkAssembly)
+            val controllersAssembly = ControllersAssemblyImpl(storageAssembly, miscAssembly, servicesAssembly)
+
+            return DependencyInjection(
+                mappersAssembly,
+                storageAssembly,
+                networkAssembly,
+                miscAssembly,
+                servicesAssembly,
+                controllersAssembly
+            )
+        }
+    }
+}
