+++ sdk/src/main/java/com/qonversion/android/sdk/internal/di/network/NetworkAssemblyImpl.kt
@@ -1,5 +1,6 @@
 package com.qonversion.android.sdk.internal.di.network
 
+import androidx.annotation.VisibleForTesting
 import com.qonversion.android.sdk.internal.common.BASE_API_URL
 import com.qonversion.android.sdk.internal.di.mappers.MappersAssembly
 import com.qonversion.android.sdk.internal.di.misc.MiscAssembly
@@ -23,16 +24,6 @@ internal class NetworkAssemblyImpl(
     override val networkClient: NetworkClient
         get() = NetworkClientImpl(miscAssembly.jsonSerializer)
 
-    override fun getApiInteractor(retryPolicy: RetryPolicy): ApiInteractor {
-        return ApiInteractorImpl(
-            networkClient,
-            miscAssembly.exponentialDelayCalculator,
-            miscAssembly.internalConfig,
-            mappersAssembly.apiErrorMapper,
-            retryPolicy
-        )
-    }
-
     override val requestConfigurator: RequestConfigurator
         get() = RequestConfiguratorImpl(
             headerBuilder,
@@ -40,6 +31,12 @@ internal class NetworkAssemblyImpl(
             miscAssembly.internalConfig,
             miscAssembly.internalConfig
         )
+
+    override val apiInteractorExponential: ApiInteractor
+        get() = provideApiInteractor(RetryPolicy.Exponential())
+
+    override val apiInteractorInfinityExponential: ApiInteractor
+        get() = provideApiInteractor(RetryPolicy.InfiniteExponential())
 
     override val headerBuilder: HeaderBuilder
         get() = HeaderBuilderImpl(
@@ -49,4 +46,15 @@ internal class NetworkAssemblyImpl(
             miscAssembly.internalConfig,
             miscAssembly.internalConfig
         )
+
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    fun provideApiInteractor(retryPolicy: RetryPolicy): ApiInteractor {
+        return ApiInteractorImpl(
+            networkClient,
+            miscAssembly.exponentialDelayCalculator,
+            miscAssembly.internalConfig,
+            mappersAssembly.apiErrorMapper,
+            retryPolicy
+        )
+    }
 }
