+++ sdk/src/main/java/com/qonversion/android/sdk/internal/di/services/ServicesAssemblyImpl.kt
@@ -3,7 +3,6 @@ package com.qonversion.android.sdk.internal.di.services
 import com.qonversion.android.sdk.internal.di.mappers.MappersAssembly
 import com.qonversion.android.sdk.internal.di.network.NetworkAssembly
 import com.qonversion.android.sdk.internal.di.storage.StorageAssembly
-import com.qonversion.android.sdk.internal.networkLayer.RetryPolicy
 import com.qonversion.android.sdk.internal.user.UserService
 import com.qonversion.android.sdk.internal.user.UserServiceImpl
 import com.qonversion.android.sdk.internal.userProperties.UserPropertiesService
@@ -18,14 +17,14 @@ internal class ServicesAssemblyImpl(
     override val userPropertiesService: UserPropertiesService
         get() = UserPropertiesServiceImpl(
             networkAssembly.requestConfigurator,
-            networkAssembly.getApiInteractor(RetryPolicy.InfiniteExponential()),
+            networkAssembly.apiInteractorInfinityExponential,
             mappersAssembly.userPropertiesMapper
         )
 
     override val userService: UserService
         get() = UserServiceImpl(
             networkAssembly.requestConfigurator,
-            networkAssembly.getApiInteractor(RetryPolicy.Exponential()),
+            networkAssembly.apiInteractorExponential,
             mappersAssembly.userMapper,
             storageAssembly.sharedPreferencesStorage
         )
